+++ CHANGELOG.md
@@ -13,8 +13,8 @@ Possible log types:
 
 ## [Unreleased]
 
+- `[added]` Add `map_async` for async functions (#165)
 - `[fixed]` Add `do_async()` to handle edge case in `do()` involving multiple inlined awaits (#149)
-
 - `[added]` Add support for Python 3.12 (#157)
 
 ## [0.15.0] - 2023-12-04